Creacion de producto corregido y funcionando
-Terminada pantalla creacion de producto
-Separado materias primas e insumos de la creacion del producto y
puestos en pantallas diferentes
- Agregado orphanRemoval a las relaciones @OneToMany
