+++ src/main/java/ar/edu/utn/sigmaproject/controller/ProductCreationController.java
@@ -9,11 +9,6 @@ import javax.xml.datatype.DatatypeConfigurationException;
 import javax.xml.datatype.DatatypeFactory;
 import javax.xml.datatype.Duration;
 
-import ar.edu.utn.sigmaproject.domain.*;
-import ar.edu.utn.sigmaproject.domain.Process;
-import ar.edu.utn.sigmaproject.service.*;
-import ar.edu.utn.sigmaproject.util.RepositoryHelper;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.transaction.annotation.Transactional;
@@ -34,7 +29,6 @@ import org.zkoss.zk.ui.select.annotation.VariableResolver;
 import org.zkoss.zk.ui.select.annotation.Wire;
 import org.zkoss.zk.ui.select.annotation.WireVariable;
 import org.zkoss.zk.ui.util.Clients;
-import org.zkoss.zul.Bandbox;
 import org.zkoss.zul.Button;
 import org.zkoss.zul.Caption;
 import org.zkoss.zul.Checkbox;
@@ -50,8 +44,24 @@ import org.zkoss.zul.Selectbox;
 import org.zkoss.zul.Textbox;
 import org.zkoss.zul.Window;
 
+import ar.edu.utn.sigmaproject.domain.MeasureUnit;
+import ar.edu.utn.sigmaproject.domain.MeasureUnitType;
+import ar.edu.utn.sigmaproject.domain.Piece;
+import ar.edu.utn.sigmaproject.domain.Process;
+import ar.edu.utn.sigmaproject.domain.ProcessType;
+import ar.edu.utn.sigmaproject.domain.Product;
+import ar.edu.utn.sigmaproject.domain.ProductCategory;
+import ar.edu.utn.sigmaproject.service.MeasureUnitRepository;
+import ar.edu.utn.sigmaproject.service.MeasureUnitTypeRepository;
+import ar.edu.utn.sigmaproject.service.PieceRepository;
+import ar.edu.utn.sigmaproject.service.ProcessRepository;
+import ar.edu.utn.sigmaproject.service.ProcessTypeRepository;
+import ar.edu.utn.sigmaproject.service.ProductCategoryRepository;
+import ar.edu.utn.sigmaproject.service.ProductRepository;
+import ar.edu.utn.sigmaproject.util.RepositoryHelper;
+
 @VariableResolver(org.zkoss.zkplus.spring.DelegatingVariableResolver.class)
-public class ProductCreationController extends SelectorComposer<Component>{
+public class ProductCreationController extends SelectorComposer<Component> {
 	private static final long serialVersionUID = 1L;
 
 	final Logger logger = LoggerFactory.getLogger(ProductCreationController.class);
@@ -124,54 +134,10 @@ public class ProductCreationController extends SelectorComposer<Component>{
 	Grid productGrid;
 	@Wire
 	Combobox productCategoryCombobox;
-
-	// supply
-	@Wire
-	Button openSupplyListButton;
-	@Wire
-	Button closeSupplyListButton;
-	@Wire
-	Component supplyCreationBlock;
-	@Wire
-	Bandbox supplyTypeBandbox;
-	@Wire
-	Listbox supplyTypePopupListbox;
-	@Wire
-	Listbox supplyListbox;
-	@Wire
-	Doublebox supplyQuantityDoublebox;
-	@Wire
-	Button saveSupplyButton;
-	@Wire
-	Button resetSupplyButton;
-	@Wire
-	Button deleteSupplyButton;
-	@Wire
-	Button cancelSupplyButton;
-
-	// raw material
 	@Wire
 	Button openRawMaterialListButton;
 	@Wire
-	Button closeRawMaterialListButton;
-	@Wire
-	Component rawMaterialCreationBlock;
-	@Wire
-	Bandbox rawMaterialTypeBandbox;
-	@Wire
-	Listbox rawMaterialTypePopupListbox;
-	@Wire
-	Listbox rawMaterialListbox;
-	@Wire
-	Doublebox rawMaterialQuantityDoublebox;
-	@Wire
-	Button saveRawMaterialButton;
-	@Wire
-	Button resetRawMaterialButton;
-	@Wire
-	Button deleteRawMaterialButton;
-	@Wire
-	Button cancelRawMaterialButton;
+	Button openSupplyListButton;
 
 	// services
 	@WireVariable
@@ -183,32 +149,22 @@ public class ProductCreationController extends SelectorComposer<Component>{
 	@WireVariable
 	private ProductRepository productRepository;
 	@WireVariable
+	private PieceRepository pieceRepository;
+	@WireVariable
+	private ProcessRepository processRepository;
+	@WireVariable
 	private ProductCategoryRepository productCategoryRepository;
-	@WireVariable
-	private RawMaterialTypeRepository rawMaterialTypeRepository;
-	@WireVariable
-	private SupplyTypeRepository supplyTypeRepository;
 
 	// attributes
 	private Product currentProduct;
 	private Piece currentPiece;
 	@SuppressWarnings("rawtypes")
 	private EventQueue eq;
-	private Supply currentSupply;
-	private SupplyType currentSupplyType;
-	private RawMaterial currentRawMaterial;
-	private RawMaterialType currentRawMaterialType;
 
 	// list
 	private List<Piece> pieceList;
 	private List<ProcessType> processTypeList;
 	private List<Process> listboxProcessList;
-	private List<Supply> supplyList;
-	private List<Supply> lateDeleteSupplyList;
-	private List<SupplyType> supplyTypePopupList;
-	private List<RawMaterial> rawMaterialList;
-	private List<RawMaterial> lateDeleteRawMaterialList;
-	private List<RawMaterialType> rawMaterialTypePopupList;
 
 	// list models
 	private ListModelList<Piece> pieceListModel;
@@ -217,16 +173,16 @@ public class ProductCreationController extends SelectorComposer<Component>{
 	private ListModelList<MeasureUnit> lengthMeasureUnitListModel;
 	private ListModelList<MeasureUnit> depthMeasureUnitListModel;
 	private ListModelList<MeasureUnit> widthMeasureUnitListModel;
-	private ListModelList<Supply> supplyListModel;
-	private ListModelList<SupplyType> supplyTypePopupListModel;
-	private ListModelList<RawMaterial> rawMaterialListModel;
-	private ListModelList<RawMaterialType> rawMaterialTypePopupListModel;
 
 	@SuppressWarnings({ "unchecked", "rawtypes" })
 	@Override
 	public void doAfterCompose(Component comp) throws Exception{
 		super.doAfterCompose(comp);
 		processTypeList = processTypeRepository.findAll();
+		if(processTypeList.isEmpty()) {
+			new RepositoryHelper().generateProcessType(processTypeRepository);
+			processTypeList = processTypeRepository.findAll();
+		}
 		processTypeListModel = new ListModelList<>(processTypeList);
 		processListbox.setModel(processTypeListModel);
 		listboxProcessList = new ArrayList<>();// listbox que contiene los procesos de la pieza seleccionada o vacio si es una nueva pieza
@@ -258,21 +214,6 @@ public class ProductCreationController extends SelectorComposer<Component>{
 				fillPieceCopy(value);
 			}
 		});
-
-		lateDeleteSupplyList = new ArrayList<Supply>();
-		supplyList = new ArrayList<Supply>();
-		supplyListModel = new ListModelList<Supply>(supplyList);
-		supplyListbox.setModel(supplyListModel);
-		currentSupply = null;
-		currentSupplyType = null;
-
-		lateDeleteRawMaterialList = new ArrayList<>();
-		rawMaterialList = new ArrayList<>();
-		rawMaterialListModel = new ListModelList<>(rawMaterialList);
-		rawMaterialListbox.setModel(rawMaterialListModel);
-		currentRawMaterial = null;
-		currentRawMaterialType = null;
-
 		productCategoryListModel = new ListModelList<>(productCategoryRepository.findAll());
 		productCategoryCombobox.setModel(productCategoryListModel);
 	}
@@ -323,18 +264,21 @@ public class ProductCreationController extends SelectorComposer<Component>{
 		if (image != null) {
 			currentProduct.setImageData(image.getByteData());			
 		}
+		currentProduct.setPieces(pieceList);
+		for(Piece eachPiece : currentProduct.getPieces()) {
+			for(Process eachProcess : eachPiece.getProcesses()) {
+				processRepository.save(eachProcess);
+			}
+			pieceRepository.save(eachPiece);
+		}
 		productRepository.save(currentProduct);
 
 		// mostrar mensaje al user
 		Clients.showNotification("Producto guardado");
 
-		// limpiar todo
-		currentProduct = null;
 		currentPiece = null;
 		refreshViewProduct();
 		refreshViewPiece();
-		supplyCreationBlock.setVisible(false);
-		rawMaterialCreationBlock.setVisible(false);
 	}
 
 	public void doUploadProductPhoto(org.zkoss.image.AImage media) {
@@ -408,7 +352,7 @@ public class ProductCreationController extends SelectorComposer<Component>{
 				return;
 			}
 		}
-		String pieceName = pieceNameTextbox.getText();
+		String pieceName = pieceNameTextbox.getText().toUpperCase();
 		MeasureUnit lengthMeasureUnit = null;
 		MeasureUnit depthMeasureUnit = null;
 		MeasureUnit widthMeasureUnit = null;
@@ -442,7 +386,6 @@ public class ProductCreationController extends SelectorComposer<Component>{
 			currentPiece.setSize(pieceSize);
 			currentPiece.setUnits(pieceUnits);
 			currentPiece.setGroup(pieceIsGroup);
-			//TODO corroborar si al modificar tbn se modifica adentro de la lista
 		}
 		currentPiece.setProcesses(listboxProcessList);
 		// actualizamos el view
@@ -467,8 +410,9 @@ public class ProductCreationController extends SelectorComposer<Component>{
 			pieceList = new ArrayList<Piece>();
 			pieceListModel = new ListModelList<Piece>(pieceList);
 			pieceListbox.setModel(pieceListModel);
-			supplyList = new ArrayList<Supply>();
-			rawMaterialList = new ArrayList<RawMaterial>();
+			// no se permite agregar insumos o matertias prima mientras el producto no esté guardado
+			openRawMaterialListButton.setDisabled(true);
+			openSupplyListButton.setDisabled(true);
 		} else {
 			productCaption.setLabel("Edicion de Producto");
 			deleteProductButton.setDisabled(false);
@@ -501,13 +445,9 @@ public class ProductCreationController extends SelectorComposer<Component>{
 			pieceList = currentProduct.getPieces();
 			pieceListModel = new ListModelList<>(pieceList);
 			pieceListbox.setModel(pieceListModel);
-			supplyList = currentProduct.getSupplies();
-			rawMaterialList = currentProduct.getRawMaterials();
+			openRawMaterialListButton.setDisabled(false);
+			openSupplyListButton.setDisabled(false);
 		}
-		refreshViewSupply();
-		refreshSupplyTypePopup();
-		refreshViewRawMaterial();
-		refreshRawMaterialTypePopup();
 	}
 
 	private void refreshViewPiece() {
@@ -759,7 +699,6 @@ public class ProductCreationController extends SelectorComposer<Component>{
 			Messagebox.show("Esta seguro que quiere eliminar el producto? Se eliminaran las piezas y procesos relacionados", "Confirmar Eliminacion", Messagebox.OK | Messagebox.CANCEL, Messagebox.QUESTION, new org.zkoss.zk.ui.event.EventListener() {
 				public void onEvent(Event evt) throws InterruptedException {
 					if (evt.getName().equals("onOK")) {
-						// la eliminacion de las piezas y procesos relacionados al producto se realizan en el servicio
 						productRepository.delete(currentProduct);
 						currentProduct = null;
 						currentPiece = null;
@@ -806,6 +745,7 @@ public class ProductCreationController extends SelectorComposer<Component>{
 		pieceCreationBlock.setVisible(true);
 		processCreationBlock.setVisible(true);
 	}
+	
 	@Listen("onClick = #pieceCopyButton")
 	public void doPieceCopyButtonClick() {
 		newPieceButtonClick();
@@ -852,293 +792,20 @@ public class ProductCreationController extends SelectorComposer<Component>{
 		}
 		refreshViewProcess();
 	}
-
-	// supply methods
-
-	@Listen("onClick = #openSupplyListButton")
-	public void showSupplyCreationBlock() {
-		supplyCreationBlock.setVisible(true);
-	}
-
-	@Listen("onClick = #closeSupplyListButton")
-	public void hideSupplyCreationBlock() {
-		supplyCreationBlock.setVisible(false);
-	}
-
-	private void refreshViewSupply() {
-		if (currentSupply == null) {
-			// borramos el text del insumo
-			// deseleccionamos la tabla y borramos la cantidad
-			supplyTypeBandbox.setDisabled(false);
-			supplyTypeBandbox.setValue("");
-			supplyQuantityDoublebox.setValue(null);
-			currentSupplyType = null;
-			deleteSupplyButton.setDisabled(true);
-			cancelSupplyButton.setDisabled(true);
-		} else {
-			currentSupplyType = currentSupply.getSupplyType();
-			supplyTypeBandbox.setDisabled(true);// no se permite modificar en la edicion
-			supplyTypeBandbox.setValue(currentSupplyType.getDescription());
-			supplyQuantityDoublebox.setValue(currentSupply.getQuantity().doubleValue());
-			deleteSupplyButton.setDisabled(false);
-			cancelSupplyButton.setDisabled(false);
-		}
-		supplyTypePopupListbox.clearSelection();
-		refreshSupplyListbox();
-	}
-
-	private void refreshSupplyListbox() {
-		supplyListModel = new ListModelList<Supply>(supplyList);
-		supplyListbox.setModel(supplyListModel);
-	}
-
-	private void refreshSupplyTypePopup() {// el popup se actualiza en base a la lista
-		supplyTypePopupList = supplyTypeRepository.findAll();
-		for(Supply supply : supplyList) {
-			supplyTypePopupList.remove(supply.getSupplyType());// sacamos del popup
-		}
-		supplyTypePopupListModel = new ListModelList<SupplyType>(supplyTypePopupList);
-		supplyTypePopupListbox.setModel(supplyTypePopupListModel);
-	}
-
-	@Listen("onSelect = #supplyTypePopupListbox")
-	public void selectionSupplyTypePopupListbox() {
-		currentSupplyType = supplyTypePopupListbox.getSelectedItem().getValue();
-		supplyTypeBandbox.setValue(currentSupplyType.getDescription());
-		supplyTypeBandbox.close();
-	}
-
-	@Listen("onSelect = #supplyListbox")
-	public void selectSupply() {
-		if(supplyListModel.isSelectionEmpty()){
-			//just in case for the no selection
-			currentSupply = null;
-		} else {
-			if(currentSupply == null) {// permite la seleccion solo si no existe nada seleccionado
-				currentSupply = supplyListbox.getSelectedItem().getValue();
-				currentSupplyType = currentSupply.getSupplyType();
-				refreshViewSupply();
-			}
-		}
-		supplyListModel.clearSelection();
-	}
-
-	@Listen("onClick = #cancelSupplyButton")
-	public void cancelSupply() {
-		currentSupply = null;
-		refreshViewSupply();
-	}
-
-	@Listen("onClick = #resetSupplyButton")
-	public void resetSupply() {
-		refreshViewSupply();
-	}
-
-	@SuppressWarnings({ "unchecked", "rawtypes" })
-	@Listen("onClick = #deleteSupplyButton")
-	public void deleteSupply() {
-		if(currentSupply != null) {
-			Messagebox.show("Esta seguro que desea eliminar " + currentSupply.getSupplyType().getDescription() + "?", "Confirmar Eliminacion", Messagebox.OK | Messagebox.CANCEL, Messagebox.QUESTION, new org.zkoss.zk.ui.event.EventListener() {
-				public void onEvent(Event evt) throws InterruptedException {
-					if (evt.getName().equals("onOK")) {
-						if(currentSupply.getId() != null) {// si el detalle existe en la bd
-							lateDeleteSupplyList.add(currentSupply);// agregamos a la lista de eliminacion por las dudas que se vuelva a agregar
-						}
-						supplyList.remove(currentSupply);// quitamos de la lista
-						currentSupply = null;// eliminamos
-						refreshSupplyTypePopup();// actualizamos el popup para que aparezca el insumo eliminado
-						refreshViewSupply();
-					}
-				}
-			});
-		} 
-	}
-
-	@Listen("onClick = #saveSupplyButton")
-	public void saveSupply() {
-		if(supplyQuantityDoublebox.getValue()==null || supplyQuantityDoublebox.getValue()<=0) {
-			Clients.showNotification("Ingresar Cantidad del Insumo", supplyQuantityDoublebox);
-			return;
-		}
-		if(currentSupplyType == null) {
-			Clients.showNotification("Debe seleccionar un Insumo", supplyTypeBandbox);
-			return;
-		}
-		// buscamos si no esta eliminado
-		Supply aux = null;
-		for(Supply lateDeleteSupply : lateDeleteSupplyList) {
-			if(currentSupplyType.getId().equals(lateDeleteSupply.getId())) {
-				aux = lateDeleteSupply;
-			}
-		}
-		if(aux != null) {
-			lateDeleteSupplyList.remove(aux);// lo eliminamos de la lista de eliminacion tardia porque el sera agregado nuevamente
-		}
-		double supplyQuantity = supplyQuantityDoublebox.getValue();
-		if(currentSupply == null) { // es nuevo
-			if(aux != null) {// si ya estaba en un detalle
-				aux.setQuantity(BigDecimal.valueOf(supplyQuantity));;
-				supplyList.add(aux);
-			} else {
-				// se crea un detalle sin id porque recien se le asignara uno al momento de grabarse definitivamente
-				currentSupply = new Supply(null, currentSupplyType, BigDecimal.valueOf(supplyQuantity));
-				supplyList.add(currentSupply);
-			}
-		} else { // se edita
-			currentSupply.setSupplyType(currentSupplyType);;
-			currentSupply.setQuantity(BigDecimal.valueOf(supplyQuantity));
-		}
-		refreshSupplyTypePopup();// actualizamos el popup
-		currentSupply = null;
-		refreshViewSupply();
-	}
-
-	// raw material methods
-
+	
 	@Listen("onClick = #openRawMaterialListButton")
-	public void showRawMaterialCreationBlock() {
-		rawMaterialCreationBlock.setVisible(true);
+	public void openRawMaterialListButtonClick() {
+		Executions.getCurrent().setAttribute("selected_product", currentProduct);
+		Window window = (Window)Executions.createComponents(
+				"/product_raw_material.zul", null, null);
+		window.doModal();
 	}
-
-	@Listen("onClick = #closeRawMaterialListButton")
-	public void hideRawMaterialCreationBlock() {
-		rawMaterialCreationBlock.setVisible(false);
+	
+	@Listen("onClick = #openSupplyListButton")
+	public void openSupplyListButtonClick() {
+		Executions.getCurrent().setAttribute("selected_product", currentProduct);
+		Window window = (Window)Executions.createComponents(
+				"/product_supply.zul", null, null);
+		window.doModal();
 	}
-
-	private void refreshViewRawMaterial() {
-		if (currentRawMaterial == null) {
-			// borramos el text de la materia prima
-			// deseleccionamos la tabla y borramos la cantidad
-			rawMaterialTypeBandbox.setDisabled(false);
-			rawMaterialTypeBandbox.setValue("");
-			rawMaterialQuantityDoublebox.setValue(null);
-			currentRawMaterialType = null;
-			deleteRawMaterialButton.setDisabled(true);
-			cancelRawMaterialButton.setDisabled(true);
-		} else {
-			currentRawMaterialType = currentRawMaterial.getRawMaterialType();
-			rawMaterialTypeBandbox.setDisabled(true);// no se permite modificar en la edicion
-			rawMaterialTypeBandbox.setValue(currentRawMaterialType.getName());
-			rawMaterialQuantityDoublebox.setValue(currentRawMaterial.getQuantity().doubleValue());
-			deleteRawMaterialButton.setDisabled(false);
-			cancelRawMaterialButton.setDisabled(false);
-		}
-		rawMaterialTypePopupListbox.clearSelection();
-		refreshRawMaterialListbox();
-	}
-
-	private void refreshRawMaterialListbox() {
-		rawMaterialListModel = new ListModelList<RawMaterial>(rawMaterialList);
-		rawMaterialListbox.setModel(rawMaterialListModel);
-	}
-
-	private void refreshRawMaterialTypePopup() {// el popup se actualiza en base a la lista
-		rawMaterialTypePopupList = rawMaterialTypeRepository.findAll();
-		for(RawMaterial rawMaterial : rawMaterialList) {
-			rawMaterialTypePopupList.remove(rawMaterial.getRawMaterialType());// sacamos del popup
-		}
-		rawMaterialTypePopupListModel = new ListModelList<>(rawMaterialTypePopupList);
-		rawMaterialTypePopupListbox.setModel(rawMaterialTypePopupListModel);
-	}
-
-	@Listen("onSelect = #rawMaterialTypePopupListbox")
-	public void selectionRawMaterialTypePopupListbox() {
-		currentRawMaterialType = (RawMaterialType) rawMaterialTypePopupListbox.getSelectedItem().getValue();
-		rawMaterialTypeBandbox.setValue(currentRawMaterialType.getName());
-		rawMaterialTypeBandbox.close();
-	}
-
-	@Listen("onSelect = #rawMaterialListbox")
-	public void selectRawMaterial() {
-		if(rawMaterialListModel.isSelectionEmpty()){
-			//just in case for the no selection
-			currentRawMaterial = null;
-		} else {
-			if(currentRawMaterial == null) {// permite la seleccion solo si no existe nada seleccionado
-				currentRawMaterial = rawMaterialListbox.getSelectedItem().getValue();
-				currentRawMaterialType = currentRawMaterial.getRawMaterialType();
-				refreshViewRawMaterial();
-			}
-		}
-		rawMaterialListModel.clearSelection();
-	}
-
-	@Listen("onClick = #cancelRawMaterialButton")
-	public void cancelRawMaterial() {
-		currentRawMaterial = null;
-		refreshViewRawMaterial();
-	}
-
-	@Listen("onClick = #resetRawMaterialButton")
-	public void resetRawMaterial() {
-		refreshViewRawMaterial();
-	}
-
-	@SuppressWarnings({ "unchecked", "rawtypes" })
-	@Listen("onClick = #deleteRawMaterialButton")
-	public void deleteRawMaterial() {
-		if(currentRawMaterial != null) {
-			Messagebox.show("Esta seguro que desea eliminar " + currentRawMaterial.getRawMaterialType().getName() + "?", "Confirmar Eliminacion", Messagebox.OK | Messagebox.CANCEL, Messagebox.QUESTION, new org.zkoss.zk.ui.event.EventListener() {
-				public void onEvent(Event evt) throws InterruptedException {
-					if (evt.getName().equals("onOK")) {
-						if(currentRawMaterial.getId() != null) {// si el detalle existe en la bd
-							lateDeleteRawMaterialList.add(currentRawMaterial);// agregamos a la lista de eliminacion por las dudas que se vuelva a agregar
-						}
-						rawMaterialList.remove(currentRawMaterial);// quitamos de la lista
-						currentRawMaterial = null;// eliminamos
-						refreshRawMaterialTypePopup();// actualizamos el popup para que aparezca vuelva a aparecer el eliminado
-						refreshViewRawMaterial();
-					}
-				}
-			});
-		} 
-	}
-
-	@Listen("onClick = #saveRawMaterialButton")
-	public void saveRawMaterial() {
-		if(rawMaterialQuantityDoublebox.getValue()==null || rawMaterialQuantityDoublebox.getValue()<=0) {
-			Clients.showNotification("Ingresar Cantidad de la Materia Prima", rawMaterialQuantityDoublebox);
-			return;
-		}
-		if(currentRawMaterialType == null) {
-			Clients.showNotification("Debe seleccionar una Materia Prima", rawMaterialTypeBandbox);
-			return;
-		}
-		// buscamos si no esta eliminado
-		RawMaterial aux = null;
-		for(RawMaterial lateDeleteRawMaterial : lateDeleteRawMaterialList) {
-			if(currentRawMaterialType.getId().equals(lateDeleteRawMaterial.getId())) {
-				aux = lateDeleteRawMaterial;
-			}
-		}
-		if(aux != null) {
-			lateDeleteRawMaterialList.remove(aux);// lo eliminamos de la lista de eliminacion tardia porque el sera agregado nuevamente
-		}
-		double rawMaterialQuantity = rawMaterialQuantityDoublebox.getValue();
-		if(currentRawMaterial == null) { // es nuevo
-			if(aux != null) {// si ya estaba en un detalle
-				aux.setQuantity(BigDecimal.valueOf(rawMaterialQuantity));;
-				rawMaterialList.add(aux);
-			} else {
-				// se crea un detalle sin id porque recien se le asignara uno al momento de grabarse definitivamente
-				currentRawMaterial = new RawMaterial(currentRawMaterialType, BigDecimal.valueOf(rawMaterialQuantity));
-				rawMaterialList.add(currentRawMaterial);
-			}
-		} else { // se edita
-			currentRawMaterial.setRawMaterialType(currentRawMaterialType);;
-			currentRawMaterial.setQuantity(BigDecimal.valueOf(rawMaterialQuantity));
-		}
-		refreshRawMaterialTypePopup();// actualizamos el popup
-		currentRawMaterial = null;
-		refreshViewRawMaterial();
-	}
-
-	public String getMeasureUnitName(MeasureUnit measureUnit) {
-		if (measureUnit != null) {
-			return measureUnit.getName();
-		} else {
-			return "[Sin Unidad de Medida]";
-		}
-	}
-
 }
