+++ src/main/java/ar/edu/utn/sigmaproject/domain/Supply.java
@@ -13,9 +13,6 @@ public class Supply implements Serializable, Cloneable {
 	Long id;
 
 	@ManyToOne
-	Product product;
-
-	@ManyToOne
 	SupplyType supplyType;
 
 	BigDecimal quantity = BigDecimal.ZERO;
@@ -25,8 +22,7 @@ public class Supply implements Serializable, Cloneable {
 
 	}
 
-	public Supply(Product product, SupplyType supplyType, BigDecimal quantity) {
-		this.product = product;
+	public Supply(SupplyType supplyType, BigDecimal quantity) {
 		this.supplyType = supplyType;
 		this.quantity = quantity;
 		isClone = false;
@@ -38,14 +34,6 @@ public class Supply implements Serializable, Cloneable {
 
 	public void setId(Long id) {
 		this.id = id;
-	}
-
-	public Product getProduct() {
-		return product;
-	}
-
-	public void setProduct(Product product) {
-		this.product = product;
 	}
 
 	public SupplyType getSupplyType() {
