+++ src/main/java/ar/edu/utn/sigmaproject/service/SupplyRepository.java
@@ -1,0 +1,10 @@
+package ar.edu.utn.sigmaproject.service;
+
+import org.springframework.stereotype.Repository;
+
+import ar.edu.utn.sigmaproject.domain.Supply;
+
+@Repository
+public interface SupplyRepository extends SearchableRepository<Supply, Long> {
+
+}
